feat(flow-editor): round-trip fidelity — extract pure robotTextIO + fix comment corruption
Story: Flow Editor — Verification & Hardening (P0 round-trip).

Extract the RF text parse/serialize path out of RobotEditor.vue (3486 LOC) into
a pure, unit-tested module `robotTextIO.ts` — single source of truth, no more
duplicate logic that drifts (the failure mode the RobotEditorEscapeRoundTrip
test had to guard against by re-implementing helpers verbatim).

Fixes two silent-corruption bugs surfaced by the new golden-corpus test:
- Inline trailing comments (`Click  ${sel}  # note`) were swallowed as an arg
  and escaped to `\# note`. Now preserved on `RobotStep.trailingComment` and
  re-emitted verbatim.
- A column-0 comment inside `*** Test Cases ***` / `*** Keywords ***` became an
  item literally NAMED `# comment`. Now buffered as `leadingComments` on the
  next item (or `form.trailingComments` if dangling) and re-emitted.

Round-trip contract documented + pinned: idempotency over a real golden corpus
(byte-stable second pass), structural-equality re-parse, and regression pins for
escapes, `${}`/`@{}`/`&{}`/`%{}` variables, `...` continuation folding, and
nested FOR/IF/ELSE/END + RETURN.

Adds the BMAD story doc capturing the party-mode decisions and the full AC/test
plan for the remaining increments (file-settings+Variables inline, libdoc-per-
environment keyword discovery, control-structure e2e).

Tests: 754 frontend vitest green (incl. 20 new FlowEditorRoundTrip), vue-tsc clean.
