+++ frontend/src/components/editor/RobotEditor.vue
@@ -23,6 +23,24 @@ import {
   hasBrowserLibraryImport,
   hasRoboScopeHealImport,
 } from '@/utils/healToggle'
+// Story: Flow Editor — Verification & Hardening. The RF text parse/serialize
+// path lives in one pure, unit-tested module (round-trip fidelity contract).
+// RobotEditor delegates to it instead of carrying duplicate logic that drifts.
+import {
+  STEP_TYPES,
+  VAR_SCOPES,
+  LOOP_FLAVORS,
+  makeStep,
+  parseRobotText,
+  serializeRobotForm,
+  type StepType,
+  type RobotStep,
+  type RobotSettingEntry,
+  type RobotVariable,
+  type RobotTestCase,
+  type RobotKeyword,
+  type RobotForm,
+} from '@/components/editor/robotTextIO'
 
 // CodeMirror imports
 import { EditorView, keymap, lineNumbers, highlightActiveLine, highlightSpecialChars } from '@codemirror/view'
@@ -44,67 +62,8 @@ const emit = defineEmits<{
 
 const { t } = useI18n()
 
-// --- Step Types ---
-type StepType = 'keyword' | 'assignment' | 'var' | 'for' | 'end' | 'if' | 'else_if' | 'else'
-  | 'while' | 'try' | 'except' | 'finally' | 'break' | 'continue' | 'return' | 'comment'
-
-const STEP_TYPES: StepType[] = [
-  'keyword', 'assignment', 'var', 'comment',
-  'for', 'if', 'else_if', 'else', 'while',
-  'try', 'except', 'finally',
-  'end', 'break', 'continue', 'return',
-]
-
-const VAR_SCOPES = ['LOCAL', 'TEST', 'TASK', 'SUITE', 'GLOBAL'] as const
-
-const LOOP_FLAVORS = ['IN', 'IN RANGE', 'IN ENUMERATE', 'IN ZIP']
-
-// --- Interfaces ---
-interface RobotSettingEntry { key: string; value: string; args: string[] }
-interface RobotVariable { name: string; value: string }
-
-interface RobotStep {
-  type: StepType
-  keyword: string        // keyword/assignment: the keyword name
-  args: string[]         // keyword/assignment/return: arguments
-  returnVars: string[]   // assignment: ['${var1}', '${var2}']
-  condition: string      // if/else_if/while: condition expression
-  loopVar: string        // for: loop variable e.g. ${item}
-  loopFlavor: string     // for: IN | IN RANGE | IN ENUMERATE | IN ZIP
-  loopValues: string[]   // for: iteration values
-  exceptPattern: string  // except: error pattern
-  exceptVar: string      // except: AS ${var}
-  varScope: string       // var: scope (LOCAL|TEST|TASK|SUITE|GLOBAL)
-  comment: string        // comment: text
-  rbs_id?: string        // RECORDER-IDMAP — see flow/flowConverter.ts
-  _lineNumber?: number   // DEBUG-3 — 1-based source line; set by parseRobotToForm
-}
-
-interface RobotTestCase {
-  name: string; documentation: string; tags: string[]
-  setup: string; teardown: string; timeout: string; template: string
-  steps: RobotStep[]
-}
-interface RobotKeyword {
-  name: string; documentation: string; arguments: string[]; tags: string[]
-  setup: string; teardown: string; timeout: string; returnValue: string
-  steps: RobotStep[]
-}
-interface RobotForm {
-  settings: RobotSettingEntry[]
-  variables: RobotVariable[]
-  testCases: RobotTestCase[]
-  keywords: RobotKeyword[]
-  preambleLines: string[]
-}
-
-function makeStep(type: StepType = 'keyword'): RobotStep {
-  return {
-    type, keyword: '', args: [], returnVars: [],
-    condition: '', loopVar: '${item}', loopFlavor: 'IN', loopValues: [],
-    exceptPattern: '', exceptVar: '', varScope: '', comment: '',
-  }
-}
+// Step types, RF interfaces, constants and makeStep now live in
+// `@/components/editor/robotTextIO` (imported above) — single source of truth.
 
 // --- State ---
 // Story EDITOR-5 — Flow is the leftmost / default tab in the explorer.
@@ -266,446 +225,21 @@ const isResource = computed(() => {
 const testCaseCount = computed(() => form.testCases.length)
 const keywordCount = computed(() => form.keywords.length)
 
-// --- Step line parser ---
-const _RBS_ID_CELL = /^# rbs:([a-f0-9]{8,32})$/
+// Step line parser + step serializer now live in
+// `@/components/editor/robotTextIO` (parseStepLine / serializeStep, with the
+// escape + inline-comment + continuation round-trip fixes). See that module.
 
-/**
- * Reverse of `escapeRfToken` (defined below). Strips the leading `\`
- * from `\#…` so the user-visible value stored on `step.args[i]`
- * matches the logical selector / arg the user typed (or the recorder
- * synthesised). Without this, a `\#login-form` from disk would never
- * compare-equal to the picker's `#login-form` candidate, and the
- * selector picker would mis-flag every escaped CSS-ID as "custom".
- *
- * Currently handles only the leading `\#` case — the only escape the
- * emitter applies. If the backend emit grows new escape sequences,
- * extend here AND in `escapeRfToken`.
- */
-function unescapeRfToken(s: string): string {
-  if (s.startsWith('\\#')) return s.slice(1)
-  return s
-}
-
-function parseStepLine(raw: string): RobotStep {
-  const step = makeStep()
-  const trimmed = raw.trim()
-  if (!trimmed) return step
-
-  // Comment
-  if (trimmed.startsWith('#')) {
-    step.type = 'comment'
-    step.comment = trimmed
-    return step
-  }
-
-  // Split into cells on 2+ spaces or tab
-  const cells = trimmed.split(/  +|\t+/).filter(c => c !== '')
-  if (cells.length === 0) return step
-
-  // Story RECORDER-IDMAP — peel off the trailing `# rbs:<id>` cell
-  // (emitted by the recorder) BEFORE the rest of the parser treats
-  // it as an arg. The id stays on `step.rbs_id`; serializeStep
-  // re-appends it so save round-trips don't drop the link to the
-  // sidecar's selector group.
-  const lastCell = cells[cells.length - 1]
-  const idMatch = _RBS_ID_CELL.exec(lastCell)
-  if (idMatch) {
-    step.rbs_id = idMatch[1]
-    cells.pop()
-    if (cells.length === 0) return step
-  }
-
-  const first = cells[0]
-
-  // Control flow markers
-  if (first === 'FOR') {
-    step.type = 'for'
-    step.loopVar = cells[1] || '${item}'
-    // Find the flavor: cells[2] could be 'IN', 'IN RANGE', etc.
-    // Since we split on 2+ spaces, 'IN RANGE' stays as one cell (single space)
-    const flavor = cells[2] || 'IN'
-    step.loopFlavor = LOOP_FLAVORS.includes(flavor) ? flavor : 'IN'
-    step.loopValues = cells.slice(3)
-    return step
-  }
-  if (first === 'END') { step.type = 'end'; return step }
-  if (first === 'IF') {
-    step.type = 'if'
-    step.condition = cells.slice(1).join('    ')
-    return step
-  }
-  if (first === 'ELSE IF') {
-    step.type = 'else_if'
-    step.condition = cells.slice(1).join('    ')
-    return step
-  }
-  if (first === 'ELSE') { step.type = 'else'; return step }
-  if (first === 'WHILE') {
-    step.type = 'while'
-    step.condition = cells.slice(1).join('    ')
-    return step
-  }
-  if (first === 'TRY') { step.type = 'try'; return step }
-  if (first === 'EXCEPT') {
-    step.type = 'except'
-    const asIdx = cells.indexOf('AS')
-    if (asIdx > 1) {
-      step.exceptPattern = cells.slice(1, asIdx).join('    ')
-      step.exceptVar = cells[asIdx + 1] || ''
-    } else {
-      step.exceptPattern = cells.slice(1).join('    ')
-    }
-    return step
-  }
-  if (first === 'FINALLY') { step.type = 'finally'; return step }
-  if (first === 'BREAK') { step.type = 'break'; return step }
-  if (first === 'CONTINUE') { step.type = 'continue'; return step }
-  if (first === 'RETURN') {
-    step.type = 'return'
-    step.args = cells.slice(1)
-    return step
-  }
-  if (first === 'VAR') {
-    step.type = 'var'
-    step.returnVars = cells[1] ? [cells[1]] : ['${var}']
-    // Remaining cells are values, but check for scope= at the end
-    const rest = cells.slice(2)
-    const scopeIdx = rest.findIndex(c => /^scope=/i.test(c))
-    if (scopeIdx >= 0) {
-      step.varScope = rest[scopeIdx].replace(/^scope=/i, '')
-      step.args = rest.slice(0, scopeIdx)
-    } else {
-      step.args = rest
-    }
-    return step
-  }
-
-  // Check for variable assignment: ${var}=  Keyword  args
-  // or multi-assign: ${a}  ${b}=  Keyword  args
-  // Also handles: ${var} =  Keyword (space before =) and = as separate cell
-  const VAR_RE = /^[$@&%]\{[^}]+\}\s*=?$/
-  const returnVars: string[] = []
-  let keywordIdx = 0
-  for (let i = 0; i < cells.length; i++) {
-    const cell = cells[i].trim()
-    if (VAR_RE.test(cell)) {
-      const varName = cell.replace(/\s*=$/, '')
-      returnVars.push(varName)
-      if (cell.endsWith('=')) {
-        keywordIdx = i + 1
-        break
-      }
-      // Check if next cell is just '=' (standalone equals sign)
-      if (i + 1 < cells.length && cells[i + 1].trim() === '=') {
-        keywordIdx = i + 2
-        break
-      }
-    } else if (cell === '=' && returnVars.length > 0) {
-      // Standalone = after variables
-      keywordIdx = i + 1
-      break
-    } else {
-      keywordIdx = i
-      break
-    }
-  }
-
-  if (returnVars.length > 0 && keywordIdx < cells.length) {
-    step.type = 'assignment'
-    step.returnVars = returnVars
-    step.keyword = cells[keywordIdx]
-    // RECORDER-RF-ESCAPE — unescape `\#…` cells back to their logical
-    // form, mirroring the serializer. Without this, a fresh recording
-    // (which emits `\#login-form`) would mismatch the picker's raw
-    // `#login-form` candidate after one save round-trip.
-    step.args = cells.slice(keywordIdx + 1).map(unescapeRfToken)
-    return step
-  }
-
-  // Regular keyword call
-  step.type = 'keyword'
-  step.keyword = cells[0]
-  step.args = cells.slice(1).map(unescapeRfToken)
-  return step
-}
-
-// --- Step serializer ---
-const SEP = '    '
-
-/**
- * Mirror of `backend/src/recording/robot_emit.py::_escape_rf_token`.
- *
- * Robot Framework's lexer treats any token that STARTS with `#` as a
- * comment. After `applySelectorSwap` swaps a candidate like
- * `#login-form` into `step.args[0]` the value is the raw selector
- * (no escape). When the user saves, this serializer must add the
- * `\` so the round-tripped `.robot` keeps the selector intact —
- * otherwise the file would emit `Click    #login-form` and RF would
- * silently run Click without args.
- *
- * Idempotent: a value that's already `\#…` doesn't start with `#`
- * (it starts with `\`), so it passes through unchanged.
- */
-function escapeRfToken(s: string): string {
-  if (!s) return s
-  return s.startsWith('#') ? '\\' + s : s
-}
-
-function serializeStep(step: RobotStep): string {
-  // Story RECORDER-IDMAP — append the trailing `# rbs:<id>` after
-  // the rest of the line if the step carries an id. RF treats it as
-  // a regular line comment; the id round-trips through save → reload
-  // so the FlowEditor's selector picker stays linked to the sidecar
-  // command across reorder / insert / delete.
-  const withId = (line: string): string =>
-    step.rbs_id ? `${line}${SEP}# rbs:${step.rbs_id}` : line
-
-  // RECORDER-RF-ESCAPE — mirror the backend escape on the save path.
-  const args = step.args.map(escapeRfToken)
-
-  switch (step.type) {
-    case 'keyword':
-      return withId([step.keyword, ...args].filter(Boolean).join(SEP))
-    case 'assignment': {
-      const vars = step.returnVars.map((v, i) =>
-        i === step.returnVars.length - 1 ? v + '=' : v
-      )
-      return withId([...vars, step.keyword, ...args].filter(Boolean).join(SEP))
-    }
-    case 'for':
-      return ['FOR', step.loopVar, step.loopFlavor, ...step.loopValues].filter(Boolean).join(SEP)
-    case 'end': return 'END'
-    case 'if':
-      return ['IF', step.condition].filter(Boolean).join(SEP)
-    case 'else_if':
-      return ['ELSE IF', step.condition].filter(Boolean).join(SEP)
-    case 'else': return 'ELSE'
-    case 'while':
-      return ['WHILE', step.condition].filter(Boolean).join(SEP)
-    case 'try': return 'TRY'
-    case 'except': {
-      const parts = ['EXCEPT']
-      if (step.exceptPattern) parts.push(step.exceptPattern)
-      if (step.exceptVar) parts.push('AS', step.exceptVar)
-      return parts.join(SEP)
-    }
-    case 'finally': return 'FINALLY'
-    case 'break': return 'BREAK'
-    case 'continue': return 'CONTINUE'
-    case 'return':
-      return ['RETURN', ...step.args].filter(Boolean).join(SEP)
-    case 'var': {
-      const parts = ['VAR', step.returnVars[0] || '${var}', ...step.args]
-      if (step.varScope) parts.push('scope=' + step.varScope)
-      return parts.filter(Boolean).join(SEP)
-    }
-    case 'comment':
-      return step.comment || '# '
-    default:
-      return ''
-  }
-}
-
-// --- Main Parser ---
-const SECTION_HEADER_RE = /^\*{3}\s*(Settings?|Variables?|Test Cases?|Tasks?|Keywords?)\s*\*{0,3}/i
-
+// --- Main Parser (delegates to robotTextIO) ---
 function parseRobotToForm(content: string): boolean {
   try {
     setupSteps.clear() // Clear cached virtual steps
-    const lines = content.split('\n')
-    const newForm: RobotForm = {
-      settings: [], variables: [], testCases: [], keywords: [], preambleLines: [],
-    }
-
-    let currentSection: 'none' | 'settings' | 'variables' | 'testcases' | 'keywords' = 'none'
-    let currentItem: RobotTestCase | RobotKeyword | null = null
-    let currentItemType: 'testcase' | 'keyword' | null = null
-
-    for (let i = 0; i < lines.length; i++) {
-      const line = lines[i]
-      const trimmed = line.trimEnd()
-
-      // Section header
-      const headerMatch = trimmed.match(SECTION_HEADER_RE)
-      if (headerMatch) {
-        if (currentItem && currentItemType) {
-          if (currentItemType === 'testcase') newForm.testCases.push(currentItem as RobotTestCase)
-          else newForm.keywords.push(currentItem as RobotKeyword)
-          currentItem = null
-          currentItemType = null
-        }
-        const sectionName = headerMatch[1].toLowerCase()
-        if (sectionName.startsWith('setting')) currentSection = 'settings'
-        else if (sectionName.startsWith('variable')) currentSection = 'variables'
-        else if (sectionName.startsWith('test') || sectionName.startsWith('task')) currentSection = 'testcases'
-        else if (sectionName.startsWith('keyword')) currentSection = 'keywords'
-        continue
-      }
-
-      if (currentSection === 'none') {
-        newForm.preambleLines.push(trimmed)
-        continue
-      }
-
-      if (currentSection === 'settings') {
-        if (!trimmed || trimmed.startsWith('#')) {
-          if (trimmed.startsWith('#')) {
-            newForm.settings.push({ key: '#', value: trimmed, args: [] })
-          }
-          continue
-        }
-        // Continuation `...` — fold into the previous setting's
-        // value (newline-separated for Documentation; space-separated
-        // would also be valid but newlines preserve readability in
-        // the textarea). Without this, a multi-line `Documentation`
-        // produced N rows with `...` as the key.
-        if (trimmed.startsWith('...')) {
-          const cont = trimmed.slice(3).trim()
-          const last = newForm.settings[newForm.settings.length - 1]
-          if (last && last.key !== '#') {
-            last.value = last.value ? `${last.value}\n${cont}` : cont
-          }
-          continue
-        }
-        const parts = trimmed.split(/  +|\t+/).filter(p => p !== '')
-        if (parts.length >= 1) {
-          newForm.settings.push({ key: parts[0], value: parts[1] || '', args: parts.slice(2) })
-        }
-        continue
-      }
-
-      if (currentSection === 'variables') {
-        if (!trimmed || trimmed.startsWith('#')) {
-          if (trimmed.startsWith('#')) {
-            newForm.variables.push({ name: '#', value: trimmed })
-          }
-          continue
-        }
-        // Continuation: append to the previous variable's value.
-        if (trimmed.startsWith('...')) {
-          const cont = trimmed.slice(3).trim()
-          const last = newForm.variables[newForm.variables.length - 1]
-          if (last && last.name !== '#') {
-            last.value = last.value ? `${last.value}${SEP}${cont}` : cont
-          }
-          continue
-        }
-        const parts = trimmed.split(/  +|\t+/).filter(p => p !== '')
-        if (parts.length >= 1) {
-          newForm.variables.push({ name: parts[0], value: parts.slice(1).join(SEP) })
-        }
-        continue
-      }
-
-      if (currentSection === 'testcases' || currentSection === 'keywords') {
-        if (!trimmed) continue
-
-        const isIndented = /^\s/.test(line) || line.startsWith('\t')
-
-        if (!isIndented && trimmed) {
-          if (currentItem && currentItemType) {
-            if (currentItemType === 'testcase') newForm.testCases.push(currentItem as RobotTestCase)
-            else newForm.keywords.push(currentItem as RobotKeyword)
-          }
-          if (currentSection === 'testcases') {
-            currentItem = {
-              name: trimmed, documentation: '', tags: [],
-              setup: '', teardown: '', timeout: '', template: '', steps: [],
-            }
-            currentItemType = 'testcase'
-          } else {
-            currentItem = {
-              name: trimmed, documentation: '', arguments: [], tags: [],
-              setup: '', teardown: '', timeout: '', returnValue: '', steps: [],
-            }
-            currentItemType = 'keyword'
-          }
-          continue
-        }
-
-        if (currentItem && isIndented) {
-          const bodyTrimmed = trimmed.trim()
-
-          // Continuation line
-          if (bodyTrimmed.startsWith('...')) {
-            const contCells = bodyTrimmed.slice(3).trim().split(/  +|\t+/).filter(c => c !== '')
-            if (currentItem.steps.length > 0) {
-              const prev = currentItem.steps[currentItem.steps.length - 1]
-              switch (prev.type) {
-                case 'keyword':
-                case 'assignment':
-                case 'return':
-                  prev.args.push(...contCells)
-                  break
-                case 'for':
-                  prev.loopValues.push(...contCells)
-                  break
-                default:
-                  // Append to condition or treat as extra text
-                  if (prev.condition !== undefined && contCells.length) {
-                    prev.condition += SEP + contCells.join(SEP)
-                  }
-              }
-            } else if (currentItem.documentation) {
-              currentItem.documentation += '\n' + contCells.join(' ')
-            }
-            continue
-          }
-
-          // Setting tags: [Documentation], [Tags], etc.
-          const settingMatch = bodyTrimmed.match(/^\[(Documentation|Tags|Setup|Teardown|Timeout|Template|Arguments|Return)\]\s*(.*)/i)
-          if (settingMatch) {
-            const settingName = settingMatch[1].toLowerCase()
-            const settingValue = settingMatch[2]?.trim() || ''
-            switch (settingName) {
-              case 'documentation': currentItem.documentation = settingValue; break
-              case 'tags': currentItem.tags = settingValue ? settingValue.split(/  +|\t+/).map(t => t.trim()).filter(Boolean) : []; break
-              case 'setup': currentItem.setup = settingValue; break
-              case 'teardown': currentItem.teardown = settingValue; break
-              case 'timeout': currentItem.timeout = settingValue; break
-              case 'template':
-                if ('template' in currentItem) (currentItem as RobotTestCase).template = settingValue
-                break
-              case 'arguments':
-                if ('arguments' in currentItem) {
-                  (currentItem as RobotKeyword).arguments = settingValue
-                    ? settingValue.split(/  +|\t+/).filter(Boolean)
-                    : []
-                }
-                break
-              case 'return':
-                if ('returnValue' in currentItem) (currentItem as RobotKeyword).returnValue = settingValue
-                break
-            }
-            continue
-          }
-
-          // Parse structured step
-          // Story DEBUG-3 — annotate with the 1-based source line so
-          // the Flow Editor's "Run up to here" debug button can pass
-          // the right line to the backend's debug-launch endpoint.
-          // The annotation is metadata only; the serializer writes
-          // each step on its own line, so a save-and-reparse keeps
-          // the numbers in sync.
-          const parsed = parseStepLine(bodyTrimmed)
-          parsed._lineNumber = i + 1
-          currentItem.steps.push(parsed)
-        }
-        continue
-      }
-    }
-
-    // Flush last item
-    if (currentItem && currentItemType) {
-      if (currentItemType === 'testcase') newForm.testCases.push(currentItem as RobotTestCase)
-      else newForm.keywords.push(currentItem as RobotKeyword)
-    }
-
-    form.settings = newForm.settings
-    form.variables = newForm.variables
-    form.testCases = newForm.testCases
-    form.keywords = newForm.keywords
-    form.preambleLines = newForm.preambleLines
+    const parsed = parseRobotText(content)
+    form.settings = parsed.settings
+    form.variables = parsed.variables
+    form.testCases = parsed.testCases
+    form.keywords = parsed.keywords
+    form.preambleLines = parsed.preambleLines
+    form.trailingComments = parsed.trailingComments
     parseError.value = null
     return true
   } catch (e: unknown) {
@@ -766,129 +300,9 @@ async function lazyLoadKeywordArgs() {
   }
 }
 
-/**
- * Multi-line settings (`[Documentation]`, `[Tags]`, …) parse with the
- * `...` continuation marker — each successive `...` line concatenates
- * its cells onto the parent setting. Our parser folds those into one
- * string with `\n` separators (so the form-edit textarea is sane). At
- * serialize time we MUST reverse the fold: dump the first line after
- * the setting tag, then emit each remaining line as a `...` row.
- *
- * Without this, a multi-line `[Documentation]` got crammed into a
- * single output line containing literal `\n` characters. RF's lexer
- * sees those newlines as line breaks and treats every continuation
- * line as a NEW test case header at column 0 — the original test
- * loses all of its body and disappears on the next reload.
- *
- * Edge cases:
- *   - Empty string → emit nothing (caller already guards on truthy).
- *   - Single line → just `[setting]    line` (no continuation).
- *   - Multi-line → `[setting]    line0` followed by `...    lineN`.
- */
-function emitMultilineSetting(tag: string, value: string): string[] {
-  const parts = value.split('\n')
-  const out: string[] = []
-  out.push(SEP + tag + SEP + parts[0])
-  for (let i = 1; i < parts.length; i++) {
-    out.push(SEP + '...' + SEP + parts[i])
-  }
-  return out
-}
-
-// --- Main Serializer ---
+// --- Main Serializer (delegates to robotTextIO) ---
 function serializeFormToRobot(): string {
-  const lines: string[] = []
-
-  for (const pl of form.preambleLines) lines.push(pl)
-
-  if (form.settings.length > 0) {
-    if (lines.length > 0 && lines[lines.length - 1] !== '') lines.push('')
-    lines.push('*** Settings ***')
-    for (const s of form.settings) {
-      if (s.key === '#') { lines.push(s.value); continue }
-      // Settings whose value contains literal `\n` (typically
-      // Documentation continued across multiple `...` lines on disk)
-      // need to be split back out — see emitMultilineSetting() in
-      // the test-case path. Without this, the saved file ends up
-      // with raw newlines mid-cell and RF re-parses each
-      // continuation line as a new top-level setting.
-      if (s.value && s.value.includes('\n')) {
-        const parts = s.value.split('\n')
-        let firstLine = s.key + SEP + parts[0]
-        for (const a of s.args) firstLine += SEP + a
-        lines.push(firstLine)
-        for (let i = 1; i < parts.length; i++) {
-          lines.push('...' + SEP + parts[i])
-        }
-        continue
-      }
-      let line = s.key
-      if (s.value) line += SEP + s.value
-      for (const a of s.args) line += SEP + a
-      lines.push(line)
-    }
-  }
-
-  if (form.variables.length > 0) {
-    if (lines.length > 0 && lines[lines.length - 1] !== '') lines.push('')
-    lines.push('*** Variables ***')
-    for (const v of form.variables) {
-      if (v.name === '#') { lines.push(v.value); continue }
-      lines.push(v.value ? v.name + SEP + v.value : v.name)
-    }
-  }
-
-  if (!isResource.value && form.testCases.length > 0) {
-    if (lines.length > 0 && lines[lines.length - 1] !== '') lines.push('')
-    lines.push('*** Test Cases ***')
-    for (let tcIdx = 0; tcIdx < form.testCases.length; tcIdx++) {
-      const tc = form.testCases[tcIdx]
-      // Empty name → fall back to "Test Case N" (same shape the
-      // FlowEditor's tab strip already shows for unnamed entries).
-      // Without this the serializer emits a blank line as the test-
-      // case header; the parser then skips that blank line per its
-      // "non-empty + non-indented = header" rule, the steps below
-      // become orphaned, and the entire test case disappears on
-      // round-trip after a save → file-switch → reload cycle.
-      const tcName = tc.name.trim() || `Test Case ${tcIdx + 1}`
-      lines.push(tcName)
-      if (tc.documentation) lines.push(...emitMultilineSetting('[Documentation]', tc.documentation))
-      if (tc.tags.length > 0) lines.push(SEP + '[Tags]' + SEP + tc.tags.join(SEP))
-      if (tc.setup) lines.push(SEP + '[Setup]' + SEP + tc.setup)
-      if (tc.teardown) lines.push(SEP + '[Teardown]' + SEP + tc.teardown)
-      if (tc.timeout) lines.push(SEP + '[Timeout]' + SEP + tc.timeout)
-      if (tc.template) lines.push(SEP + '[Template]' + SEP + tc.template)
-      for (const step of tc.steps) lines.push(SEP + serializeStep(step))
-      lines.push('')
-    }
-  }
-
-  if (form.keywords.length > 0) {
-    if (lines.length > 0 && lines[lines.length - 1] !== '') lines.push('')
-    lines.push('*** Keywords ***')
-    for (let kwIdx = 0; kwIdx < form.keywords.length; kwIdx++) {
-      const kw = form.keywords[kwIdx]
-      // Same fallback as test cases — an empty keyword name on
-      // disk is equally invalid and would silently swallow the
-      // keyword's steps on reload.
-      const kwName = kw.name.trim() || `Keyword ${kwIdx + 1}`
-      lines.push(kwName)
-      if (kw.arguments.length) lines.push(SEP + '[Arguments]' + SEP + kw.arguments.join(SEP))
-      if (kw.documentation) lines.push(...emitMultilineSetting('[Documentation]', kw.documentation))
-      if (kw.tags.length > 0) lines.push(SEP + '[Tags]' + SEP + kw.tags.join(SEP))
-      if (kw.setup) lines.push(SEP + '[Setup]' + SEP + kw.setup)
-      if (kw.teardown) lines.push(SEP + '[Teardown]' + SEP + kw.teardown)
-      if (kw.timeout) lines.push(SEP + '[Timeout]' + SEP + kw.timeout)
-      if (kw.returnValue) lines.push(SEP + '[Return]' + SEP + kw.returnValue)
-      for (const step of kw.steps) lines.push(SEP + serializeStep(step))
-      lines.push('')
-    }
-  }
-
-  let result = lines.join('\n')
-  result = result.replace(/\n{3,}/g, '\n\n')
-  if (!result.endsWith('\n')) result += '\n'
-  return result
+  return serializeRobotForm(form, { isResource: isResource.value })
 }
 
 // --- Tab Switching ---
